# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JST Tailors Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_stitching, ready, delivered]
        - name: customerId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: orderId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, customerId, orderDate, status, totalAmount, advanceAmount, balanceAmount]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        orderDate:
          type: string
        deliveryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_stitching, ready, delivered]
        totalAmount:
          type: number
        advanceAmount:
          type: number
        balanceAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [customerId, orderDate, totalAmount]
      properties:
        customerId:
          type: integer
        orderDate:
          type: string
        deliveryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_stitching, ready, delivered]
          default: pending
        totalAmount:
          type: number
        advanceAmount:
          type: number
          default: 0
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        customerId:
          type: integer
        orderDate:
          type: string
        deliveryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_stitching, ready, delivered]
        totalAmount:
          type: number
        advanceAmount:
          type: number
        notes:
          type: ["string", "null"]
    Payment:
      type: object
      required: [id, orderId, customerId, amount, paymentDate]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        orderNumber:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [orderId, customerId, amount, paymentDate]
      properties:
        orderId:
          type: integer
        customerId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: string
...